requirements.txt
Flask
Pillow
opencv-python
numpy
pytesseract
PyPDF2
python-dotenv
groq
tensorflow
